Addresses issue with Gutenberg epubs
I don't know whether this is standard or not, but this addresses two
problems seen in books downloaded from Project Gutenberg, using an epub
of Rudyard Kipling's _Kim_ as an example.

First the adds Content with an anchor reference in the source link, eg `<content src="5623131657474122306_2226-h-0.htm.html#pgepubid00001"/>` the epub library does not handle these very well.

Secondly, the `navMap` only includes top-level items, while the parser
expects to see child items. I have adjusted the behaviour to handle
top-level or child items if they exist, this slightly changes the
`TestReader` function because it now finds several more "chapters",
which may or may not be authentic chapters in the document - this is
where my knowledge of the format remains somewhat limited, but I believe
the new behaviour is correct.